+++ imaginarium/src/common/image_diff.rs
@@ -27,8 +27,8 @@ pub fn max_pixel_diff(img1: &Image, img2: &Image) -> f64 {
     let format = img1.desc().color_format;
     let pixel_size = format.byte_count() as usize;
     let row_bytes = width * pixel_size;
-    let stride1 = img1.desc().stride as usize;
-    let stride2 = img2.desc().stride as usize;
+    let stride1 = img1.desc().stride;
+    let stride2 = img2.desc().stride;
 
     (0..height)
         .into_par_iter()
@@ -95,8 +95,8 @@ pub fn pixels_equal(img1: &Image, img2: &Image) -> bool {
     let height = img1.desc().height as usize;
     let pixel_size = img1.desc().color_format.byte_count() as usize;
     let row_bytes = width * pixel_size;
-    let stride1 = img1.desc().stride as usize;
-    let stride2 = img2.desc().stride as usize;
+    let stride1 = img1.desc().stride;
+    let stride2 = img2.desc().stride;
 
     (0..height).into_par_iter().all(|y| {
         let row1 = &img1.bytes()[y * stride1..y * stride1 + row_bytes];
