+++ imaginarium/src/ops/contrast_brightness/cpu.rs
@@ -185,7 +185,7 @@ where
 
     let width = from.desc().width as usize;
     let channels = from.desc().color_format.channel_count.channel_count() as usize;
-    let stride = from.desc().stride as usize;
+    let stride = from.desc().stride;
     let row_bytes = width * channels * size_of::<T>();
 
     let has_alpha = channels == 2 || channels == 4;
@@ -225,8 +225,8 @@ where
 #[target_feature(enable = "sse4.1")]
 unsafe fn apply_u8_gray_sse41(from: &Image, to: &mut Image, params: ContrastBrightness) {
     let width = from.desc().width as usize;
-    let in_stride = from.desc().stride as usize;
-    let out_stride = to.desc().stride as usize;
+    let in_stride = from.desc().stride;
+    let out_stride = to.desc().stride;
     let contrast = params.contrast;
     let offset = 127.5 * (1.0 - contrast) + params.brightness * 255.0;
 
@@ -310,8 +310,8 @@ unsafe fn process_row_u8_gray_sse41(
 #[target_feature(enable = "sse4.1")]
 unsafe fn apply_u8_gray_alpha_sse41(from: &Image, to: &mut Image, params: ContrastBrightness) {
     let width = from.desc().width as usize;
-    let in_stride = from.desc().stride as usize;
-    let out_stride = to.desc().stride as usize;
+    let in_stride = from.desc().stride;
+    let out_stride = to.desc().stride;
     let contrast = params.contrast;
     let offset = 127.5 * (1.0 - contrast) + params.brightness * 255.0;
 
@@ -402,8 +402,8 @@ unsafe fn process_row_u8_gray_alpha_sse41(
 #[target_feature(enable = "sse4.1")]
 unsafe fn apply_u8_rgb_sse41(from: &Image, to: &mut Image, params: ContrastBrightness) {
     let width = from.desc().width as usize;
-    let in_stride = from.desc().stride as usize;
-    let out_stride = to.desc().stride as usize;
+    let in_stride = from.desc().stride;
+    let out_stride = to.desc().stride;
     let contrast = params.contrast;
     let offset = 127.5 * (1.0 - contrast) + params.brightness * 255.0;
 
@@ -500,8 +500,8 @@ unsafe fn process_row_u8_rgb_sse41(
 #[target_feature(enable = "sse4.1")]
 unsafe fn apply_u8_rgba_sse41(from: &Image, to: &mut Image, params: ContrastBrightness) {
     let width = from.desc().width as usize;
-    let in_stride = from.desc().stride as usize;
-    let out_stride = to.desc().stride as usize;
+    let in_stride = from.desc().stride;
+    let out_stride = to.desc().stride;
     let contrast = params.contrast;
     let offset = 127.5 * (1.0 - contrast) + params.brightness * 255.0;
 
@@ -607,8 +607,8 @@ unsafe fn process_row_u8_rgba_sse41(
 #[target_feature(enable = "sse4.1")]
 unsafe fn apply_f32_gray_sse41(from: &Image, to: &mut Image, params: ContrastBrightness) {
     let width = from.desc().width as usize;
-    let in_stride = from.desc().stride as usize;
-    let out_stride = to.desc().stride as usize;
+    let in_stride = from.desc().stride;
+    let out_stride = to.desc().stride;
     let contrast = params.contrast;
     let brightness = params.brightness;
 
@@ -672,8 +672,8 @@ unsafe fn process_row_f32_gray_sse41(
 #[target_feature(enable = "sse4.1")]
 unsafe fn apply_f32_gray_alpha_sse41(from: &Image, to: &mut Image, params: ContrastBrightness) {
     let width = from.desc().width as usize;
-    let in_stride = from.desc().stride as usize;
-    let out_stride = to.desc().stride as usize;
+    let in_stride = from.desc().stride;
+    let out_stride = to.desc().stride;
     let contrast = params.contrast;
     let brightness = params.brightness;
 
@@ -747,8 +747,8 @@ unsafe fn process_row_f32_gray_alpha_sse41(
 #[target_feature(enable = "sse4.1")]
 unsafe fn apply_f32_rgb_sse41(from: &Image, to: &mut Image, params: ContrastBrightness) {
     let width = from.desc().width as usize;
-    let in_stride = from.desc().stride as usize;
-    let out_stride = to.desc().stride as usize;
+    let in_stride = from.desc().stride;
+    let out_stride = to.desc().stride;
     let contrast = params.contrast;
     let brightness = params.brightness;
 
@@ -829,8 +829,8 @@ unsafe fn process_row_f32_rgb_sse41(
 #[target_feature(enable = "sse4.1")]
 unsafe fn apply_f32_rgba_sse41(from: &Image, to: &mut Image, params: ContrastBrightness) {
     let width = from.desc().width as usize;
-    let in_stride = from.desc().stride as usize;
-    let out_stride = to.desc().stride as usize;
+    let in_stride = from.desc().stride;
+    let out_stride = to.desc().stride;
     let contrast = params.contrast;
     let brightness = params.brightness;
 
@@ -896,8 +896,8 @@ unsafe fn process_row_f32_rgba_sse41(
 #[cfg(target_arch = "aarch64")]
 unsafe fn apply_u8_gray_neon(from: &Image, to: &mut Image, params: ContrastBrightness) {
     let width = from.desc().width as usize;
-    let in_stride = from.desc().stride as usize;
-    let out_stride = to.desc().stride as usize;
+    let in_stride = from.desc().stride;
+    let out_stride = to.desc().stride;
     let contrast = params.contrast;
     let offset = 127.5 * (1.0 - contrast) + params.brightness * 255.0;
 
@@ -979,8 +979,8 @@ unsafe fn process_row_u8_gray_neon(
 #[cfg(target_arch = "aarch64")]
 unsafe fn apply_u8_gray_alpha_neon(from: &Image, to: &mut Image, params: ContrastBrightness) {
     let width = from.desc().width as usize;
-    let in_stride = from.desc().stride as usize;
-    let out_stride = to.desc().stride as usize;
+    let in_stride = from.desc().stride;
+    let out_stride = to.desc().stride;
     let contrast = params.contrast;
     let offset = 127.5 * (1.0 - contrast) + params.brightness * 255.0;
 
@@ -1060,8 +1060,8 @@ unsafe fn process_row_u8_gray_alpha_neon(
 #[cfg(target_arch = "aarch64")]
 unsafe fn apply_u8_rgb_neon(from: &Image, to: &mut Image, params: ContrastBrightness) {
     let width = from.desc().width as usize;
-    let in_stride = from.desc().stride as usize;
-    let out_stride = to.desc().stride as usize;
+    let in_stride = from.desc().stride;
+    let out_stride = to.desc().stride;
     let contrast = params.contrast;
     let offset = 127.5 * (1.0 - contrast) + params.brightness * 255.0;
 
@@ -1145,8 +1145,8 @@ unsafe fn process_row_u8_rgb_neon(
 #[cfg(target_arch = "aarch64")]
 unsafe fn apply_u8_rgba_neon(from: &Image, to: &mut Image, params: ContrastBrightness) {
     let width = from.desc().width as usize;
-    let in_stride = from.desc().stride as usize;
-    let out_stride = to.desc().stride as usize;
+    let in_stride = from.desc().stride;
+    let out_stride = to.desc().stride;
     let contrast = params.contrast;
     let offset = 127.5 * (1.0 - contrast) + params.brightness * 255.0;
 
@@ -1234,8 +1234,8 @@ unsafe fn process_row_u8_rgba_neon(
 #[cfg(target_arch = "aarch64")]
 unsafe fn apply_f32_gray_neon(from: &Image, to: &mut Image, params: ContrastBrightness) {
     let width = from.desc().width as usize;
-    let in_stride = from.desc().stride as usize;
-    let out_stride = to.desc().stride as usize;
+    let in_stride = from.desc().stride;
+    let out_stride = to.desc().stride;
     let contrast = params.contrast;
     let brightness = params.brightness;
 
@@ -1297,8 +1297,8 @@ unsafe fn process_row_f32_gray_neon(
 #[cfg(target_arch = "aarch64")]
 unsafe fn apply_f32_gray_alpha_neon(from: &Image, to: &mut Image, params: ContrastBrightness) {
     let width = from.desc().width as usize;
-    let in_stride = from.desc().stride as usize;
-    let out_stride = to.desc().stride as usize;
+    let in_stride = from.desc().stride;
+    let out_stride = to.desc().stride;
     let contrast = params.contrast;
     let brightness = params.brightness;
 
@@ -1369,8 +1369,8 @@ unsafe fn process_row_f32_gray_alpha_neon(
 #[cfg(target_arch = "aarch64")]
 unsafe fn apply_f32_rgb_neon(from: &Image, to: &mut Image, params: ContrastBrightness) {
     let width = from.desc().width as usize;
-    let in_stride = from.desc().stride as usize;
-    let out_stride = to.desc().stride as usize;
+    let in_stride = from.desc().stride;
+    let out_stride = to.desc().stride;
     let contrast = params.contrast;
     let brightness = params.brightness;
 
@@ -1444,8 +1444,8 @@ unsafe fn process_row_f32_rgb_neon(
 #[cfg(target_arch = "aarch64")]
 unsafe fn apply_f32_rgba_neon(from: &Image, to: &mut Image, params: ContrastBrightness) {
     let width = from.desc().width as usize;
-    let in_stride = from.desc().stride as usize;
-    let out_stride = to.desc().stride as usize;
+    let in_stride = from.desc().stride;
+    let out_stride = to.desc().stride;
     let contrast = params.contrast;
     let brightness = params.brightness;
